requirements.txt
argcomplete==1.10.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
compressed-rtf==1.0.6
cymem==2.0.6
Cython==0.29.14
dj-database-url==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-extensions==3.1.5
django-heroku==0.3.1
django-picklefield==3.0.1
docx2txt==0.8
ebcdic==1.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
et-xmlfile==1.1.0
extract-msg==0.28.7
gensim==3.8.3
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
IMAPClient==2.1.0
Jinja2==3.1.1
langcodes==3.3.0
lxml==4.8.0
MarkupSafe==2.1.1
murmurhash==1.0.6
mysqlclient==2.1.0
numexpr==2.8.1
numpy==1.22.3
olefile==0.46
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pathy==0.6.1
pdfminer.six==20191110
Pillow==9.0.1
preshed==3.0.6
psycopg2==2.9.3
pycryptodome==3.14.1
pydantic==1.8.2
PyMuPDF==1.19.6
pyparsing==3.0.7
PyPDF2==1.26.0
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
scipy==1.8.0
six==1.12.0
smart-open==5.2.1
sortedcontainers==2.4.0
soupsieve==2.3.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
srsly==2.4.2
stats-can==2.5.1
tables==3.7.0
text-summarizer==0.0.6
textract==1.6.5
thinc==8.0.15
tqdm==4.63.0
typer==0.4.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
wasabi==0.9.0
whitenoise==6.0.0
xlrd==1.2.0
XlsxWriter==3.0.3
